# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BharatSR",
    "description": "Scientifically Defensible Deep Learning Super-Resolution for Satellite Earth Observation (SIH26142)",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/dataset/scenes": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get Dataset Scenes",
        "description": "Paginated listing of individual scene metadata records.",
        "operationId": "get_dataset_scenes_api_dataset_scenes_get",
        "parameters": [
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(train|val|test)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Split"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetScenesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset/qa-report/{split}": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get Qa Report",
        "description": "Serve the QA verification figure (PNG) for train, val, or test split.",
        "operationId": "get_qa_report_api_dataset_qa_report__split__get",
        "parameters": [
          {
            "name": "split",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Split"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dataset/preprocessing-sample": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get Preprocessing Sample",
        "description": "Returns step-by-step visual representations of the 4 preprocessing stages:\n1. Raw DN capture\n2. Radiometric BOA Surface Reflectance [0, 1]\n3. Sub-pixel Registration & QA Cleaning\n4. 64x64 LR & 256x256 HR Training Pair",
        "operationId": "get_preprocessing_sample_api_dataset_preprocessing_sample_get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreprocessingSampleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/history/{model_name}": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get Training History",
        "description": "Return recorded epoch-by-epoch training metrics or verified final metrics for a model.",
        "operationId": "get_training_history_api_training_history__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/model-card/{model_name}": {
      "get": {
        "tags": [
          "dataset"
        ],
        "summary": "Get Model Card",
        "description": "Return architectural model card, parameter specifications, loss config, and benchmark metrics.",
        "operationId": "get_model_card_api_training_model_card__model_name__get",
        "parameters": [
          {
            "name": "model_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DatasetScenesResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneRecord"
            },
            "type": "array",
            "title": "Scenes"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "scenes"
        ],
        "title": "DatasetScenesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelCardResponse": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "parameters_count": {
            "type": "integer",
            "title": "Parameters Count"
          },
          "key_design": {
            "type": "string",
            "title": "Key Design"
          },
          "training_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Training Config"
          },
          "final_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Final Metrics"
          },
          "recommended_use": {
            "type": "string",
            "title": "Recommended Use"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model_name",
          "architecture",
          "parameters_count",
          "key_design",
          "training_config",
          "final_metrics",
          "recommended_use"
        ],
        "title": "ModelCardResponse"
      },
...
      "PreprocessingSampleResponse": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PreprocessingStageInfo"
            },
            "type": "array",
            "title": "Stages"
          },
          "training_pair": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Training Pair"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "scene_id",
          "region",
          "stages",
          "training_pair"
        ],
        "title": "PreprocessingSampleResponse"
      },
      "PreprocessingStageInfo": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "image": {
            "type": "string",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "title",
          "description",
          "image"
        ],
        "title": "PreprocessingStageInfo"
      },
...
      "SceneRecord": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "split": {
            "type": "string",
            "title": "Split"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          },
          "source_dataset": {
            "type": "string",
            "title": "Source Dataset"
          },
          "sensor": {
            "type": "string",
            "title": "Sensor"
          },
          "cloud_fraction": {
            "type": "number",
            "title": "Cloud Fraction"
          },
          "registration_rmse": {
            "type": "number",
            "title": "Registration Rmse"
          },
          "crs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crs"
          },
          "reflectance_range": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reflectance Range"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "scene_id",
          "region",
          "date",
          "split",
          "is_synthetic",
          "source_dataset",
          "sensor",
          "cloud_fraction",
          "registration_rmse"
        ],
        "title": "SceneRecord"
      },
...
      "TrainingHistoryResponse": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "is_final_only": {
            "type": "boolean",
            "title": "Is Final Only"
          },
          "epochs": {
            "items": {
              "$ref": "#/components/schemas/EpochMetric"
            },
            "type": "array",
            "title": "Epochs"
          },
          "final_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Final Metrics"
          },
          "summary_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Note"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model_name",
          "is_final_only",
          "epochs",
          "final_metrics"
        ],
        "title": "TrainingHistoryResponse"
      },
...
    }
  }
}